+++ packages/polaris-viz/src/components/shared/HorizontalBars/components/index.ts
@@ -1,3 +1,0 @@
-export {Label} from './Label';
-export {Bar} from '../../Bar';
-export {LabelWrapper} from './LabelWrapper';