+++ packages/polaris-viz/src/components/shared/HorizontalStackedBars/index.ts
@@ -1,1 +1,3 @@
 export {HorizontalStackedBars} from './HorizontalStackedBars';
+export {LabelWrapper} from '../LabelWrapper';
+export {Label} from '../Label';